+++ docs/CHANGE_LOG.md
@@ -1,5 +1,22 @@
 # JHson.js - Change Log:
 
+## Version 0.6.0:
+
+#### **Public Functions:**
+- Added "logTemplateDataWarnings()" support for "html()", which states if the template variables are not found should be logged as warnings when "write()" is called.
+
+#### **Configuration Options:**
+- Added new configuration option "variableWarningText", which states the warning text that should be shown in the console when a template variable is not used (defaults to "Template variable {{variable_name}} not found.").
+- Added new configuration option "objectErrorText", which states the error text that should be shown when an object error is detected (defaults to "Errors in object: {{error_1}}, {{error_2}}").
+
+#### **Fixes & Improvements:**
+- Fixed "clearCssFromHead()" throwing an error when clearing the style tags from the DOM documents head.
+- Updated README.md and README_NUGET.md to include the latest chained function calls.
+- Minor internal renames to make way for future enhancements.
+
+<br>
+
+
 ## Version 0.5.0:
 
 #### **Public Functions:**